# packages/contracts/swagger/spotify.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/artists": {
      "get": {
        "operationId": "ArtistsController_getAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "username",
            "required": false,
            "in": "query",
            "description": "Search by artist username",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "example": [
                  {
                    "avatar": "https://example.com/avatar.jpg",
                    "id": "1",
                    "username": "artist1",
                    "backgroundImage": "",
                    "bio": ""
                  }
                ]
              }
            },
            "description": ""
          },
          "405": {
            "description": "Method not allowed"
          },
          "408": {
            "description": "Request timeout"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Internal server error"
          },
          "501": {
            "description": "Not implemented"
          },
          "502": {
            "description": "Bad gateway"
          },
          "503": {
            "description": "Service unavailable"
          },
          "504": {
            "description": "Gateway timeout"
          },
          "505": {
            "description": "HTTP version not supported"
          },
          "507": {
            "description": "Insufficient storage"
          },
          "508": {
            "description": "Loop detected"
          },
          "default": {
            "description": ""
          }
        },
        "summary": "Get all artists with pagination",
        "tags": ["Artists"]
      },
      "post": {
        "operationId": "ArtistsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateArtistDto"
              }
            }
          }
        },
        "responses": {
          "405": {
            "description": "Method not allowed"
          },
          "408": {
            "description": "Request timeout"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Internal server error"
          },
          "501": {
            "description": "Not implemented"
          },
          "502": {
            "description": "Bad gateway"
          },
          "503": {
            "description": "Service unavailable"
          },
          "504": {
            "description": "Gateway timeout"
          },
          "505": {
            "description": "HTTP version not supported"
          },
          "507": {
            "description": "Insufficient storage"
          },
          "508": {
            "description": "Loop detected"
          },
          "default": {
            "description": ""
          }
        },
        "summary": "Create a new artist",
        "tags": ["Artists"]
      }
    },
    "/artists/{id}": {
      "get": {
        "operationId": "ArtistsController_getById",
        "parameters": [],
        "responses": {
          "405": {
            "description": "Method not allowed"
          },
          "408": {
            "description": "Request timeout"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Internal server error"
          },
          "501": {
            "description": "Not implemented"
          },
          "502": {
            "description": "Bad gateway"
          },
          "503": {
            "description": "Service unavailable"
          },
          "504": {
            "description": "Gateway timeout"
          },
          "505": {
            "description": "HTTP version not supported"
          },
          "507": {
            "description": "Insufficient storage"
          },
          "508": {
            "description": "Loop detected"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Get artist by id",
        "tags": ["Artists"]
      },
      "put": {
        "operationId": "ArtistsController_updateProfile",
        "parameters": [],
        "responses": {
          "405": {
            "description": "Method not allowed"
          },
          "408": {
            "description": "Request timeout"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Internal server error"
          },
          "501": {
            "description": "Not implemented"
          },
          "502": {
            "description": "Bad gateway"
          },
          "503": {
            "description": "Service unavailable"
          },
          "504": {
            "description": "Gateway timeout"
          },
          "505": {
            "description": "HTTP version not supported"
          },
          "507": {
            "description": "Insufficient storage"
          },
          "508": {
            "description": "Loop detected"
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Update artist profile",
        "tags": ["Artists"]
      },
      "delete": {
        "operationId": "ArtistsController_deleteProfile",
        "parameters": [],
        "responses": {
          "405": {
            "description": "Method not allowed"
          },
          "408": {
            "description": "Request timeout"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Internal server error"
          },
          "501": {
            "description": "Not implemented"
          },
          "502": {
            "description": "Bad gateway"
          },
          "503": {
            "description": "Service unavailable"
          },
          "504": {
            "description": "Gateway timeout"
          },
          "505": {
            "description": "HTTP version not supported"
          },
          "507": {
            "description": "Insufficient storage"
          },
          "508": {
            "description": "Loop detected"
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Delete artist profile",
        "tags": ["Artists"]
      }
    },
    "/artists/username/{username}": {
      "get": {
        "operationId": "ArtistsController_getByUsername",
        "parameters": [],
        "responses": {
          "405": {
            "description": "Method not allowed"
          },
          "408": {
            "description": "Request timeout"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Internal server error"
          },
          "501": {
            "description": "Not implemented"
          },
          "502": {
            "description": "Bad gateway"
          },
          "503": {
            "description": "Service unavailable"
          },
          "504": {
            "description": "Gateway timeout"
          },
          "505": {
            "description": "HTTP version not supported"
          },
          "507": {
            "description": "Insufficient storage"
          },
          "508": {
            "description": "Loop detected"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "summary": "Get artist by username",
        "tags": ["Artists"]
      }
    },
...
    "/albums": {
      "get": {
        "operationId": "AlbumsController_getAllAlbums",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Method not allowed"
          },
          "408": {
            "description": "Request timeout"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Internal server error"
          },
          "501": {
            "description": "Not implemented"
          },
          "502": {
            "description": "Bad gateway"
          },
          "503": {
            "description": "Service unavailable"
          },
          "504": {
            "description": "Gateway timeout"
          },
          "505": {
            "description": "HTTP version not supported"
          },
          "507": {
            "description": "Insufficient storage"
          },
          "508": {
            "description": "Loop detected"
          },
          "default": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all albums with pagination and filters",
        "tags": ["Albums"]
      },
      "post": {
        "operationId": "AlbumsController_createAlbum",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAlbumDto"
              }
            }
          }
        },
        "responses": {
          "405": {
            "description": "Method not allowed"
          },
          "408": {
            "description": "Request timeout"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Internal server error"
          },
          "501": {
            "description": "Not implemented"
          },
          "502": {
            "description": "Bad gateway"
          },
          "503": {
            "description": "Service unavailable"
          },
          "504": {
            "description": "Gateway timeout"
          },
          "505": {
            "description": "HTTP version not supported"
          },
          "507": {
            "description": "Insufficient storage"
          },
          "508": {
            "description": "Loop detected"
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Create a new album",
        "tags": ["Albums"]
      }
    },
    "/albums/{id}": {
      "get": {
        "operationId": "AlbumsController_getById",
        "parameters": [],
        "responses": {
          "405": {
            "description": "Method not allowed"
          },
          "408": {
            "description": "Request timeout"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Internal server error"
          },
          "501": {
            "description": "Not implemented"
          },
          "502": {
            "description": "Bad gateway"
          },
          "503": {
            "description": "Service unavailable"
          },
          "504": {
            "description": "Gateway timeout"
          },
          "505": {
            "description": "HTTP version not supported"
          },
          "507": {
            "description": "Insufficient storage"
          },
          "508": {
            "description": "Loop detected"
          },
          "default": {
            "description": ""
          }
        },
        "summary": "Get album by id",
        "tags": ["Albums"]
      },
      "put": {
        "operationId": "AlbumsController_updateAlbum",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAlbumDto"
              }
            }
          }
        },
        "responses": {
          "405": {
            "description": "Method not allowed"
          },
          "408": {
            "description": "Request timeout"
          },
          "429": {
            "description": "Too many requests"
          },
          "500": {
            "description": "Internal server error"
          },
          "501": {
            "description": "Not implemented"
          },
          "502": {
            "description": "Bad gateway"
          },
          "503": {
            "description": "Service unavailable"
          },
          "504": {
            "description": "Gateway timeout"
          },
          "505": {
            "description": "HTTP version not supported"
          },
          "507": {
            "description": "Insufficient storage"
          },
          "508": {
            "description": "Loop detected"
          },
          "default": {
            "description": ""
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Create a new album",
        "tags": ["Albums"]
      }
    }
  },
  "info": {
    "title": "@spotify/api",
    "description": "@spotify/api Swagger documentation",
    "version": "0.0.1",
    "contact": {
      "name": "Lordpluha",
      "url": "https://github.com/Lordpluha",
      "email": "vladislavteslyukofficial@gmail.com"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    },
    {
      "url": "https://spotify-clone-api-jp5z.onrender.com/",
      "description": "Remote dev server"
    }
  ],
  "components": {
    "securitySchemes": {
      "oauth2": {
        "type": "openIdConnect",
        "flows": {}
      }
    },
    "schemas": {
...
      "CreateArtistDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "User password",
            "example": "password123"
          },
          "username": {
            "type": "string",
            "description": "User username",
            "example": "user123"
          }
        },
        "required": ["email", "password", "username"]
      },
...
      "UpdateAlbumDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Playlist title"
          },
          "description": {
            "type": "string",
            "description": "",
            "example": "user123"
          }
        },
        "required": ["title"]
      },
      "CreateAlbumDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "description": "Playlist title"
          },
          "description": {
            "type": "string",
            "description": "",
            "example": "user123"
          }
        },
        "required": ["title"]
      }
    }
  },
...
}